+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+* jdbc: support reading of collections of Decimal from DB array columns
+
 # 1.5.4
 * json: ValueConverter can now be used to transform strings into types
 * json: JsonNode.getList() signature fixed